requirements.txt
langchain
langchain_core
langchain-groq
python-dotenv
arxiv
wikipedia
streamlit
langchain-community
langchain-text-splitters
pypdf
sentence-transformers
pinecone
langchain-huggingface
ddgs
langgraph